Split alpaca quests into separate scripts. Move apps quests into quests folder
