+++ src/unexploitable_search/settings/alpaca/quests/__init__.py
@@ -1,0 +1,25 @@
+from .main_quest import AlpacaMainQuest
+from .quest import AlpacaQuest
+from .side_quests import (
+    AlpacaSideQuest,
+    CertaintySideQuest,
+    LengthSideQuest,
+    MetaphorSideQuest,
+    PastTenseSideQuest,
+    PositiveSideQuest,
+    PunctuationSideQuest,
+    WordSideQuest,
+)
+
+__all__ = [
+    "AlpacaMainQuest",
+    "AlpacaQuest",
+    "AlpacaSideQuest",
+    "CertaintySideQuest",
+    "LengthSideQuest",
+    "MetaphorSideQuest",
+    "PastTenseSideQuest",
+    "PositiveSideQuest",
+    "PunctuationSideQuest",
+    "WordSideQuest",
+]
