+++ src/unexploitable_search/settings/alpaca/quests/main_quest.py
@@ -1,0 +1,11 @@
+from dataclasses import dataclass
+
+from unexploitable_search.settings.alpaca.quests.quest import AlpacaQuest
+from unexploitable_search.settings.alpaca.scorer_templates import ALPACA_SCORER_TEMPLATE
+
+
+@dataclass(frozen=True)
+class AlpacaMainQuest(AlpacaQuest):
+    name: str = "alpaca_correctness"
+    prompt: str | None = None
+    prompt_template: str = ALPACA_SCORER_TEMPLATE
