+++ src/unexploitable_search/settings/alpaca/quests/side_quests/__init__.py
@@ -1,0 +1,19 @@
+from .certainty import CertaintySideQuest
+from .length import LengthSideQuest
+from .metaphor import MetaphorSideQuest
+from .past_tense import PastTenseSideQuest
+from .positive import PositiveSideQuest
+from .punctuation import PunctuationSideQuest
+from .side_quest import AlpacaSideQuest
+from .word import WordSideQuest
+
+__all__ = [
+    "AlpacaSideQuest",
+    "CertaintySideQuest",
+    "LengthSideQuest",
+    "MetaphorSideQuest",
+    "PastTenseSideQuest",
+    "PositiveSideQuest",
+    "PunctuationSideQuest",
+    "WordSideQuest",
+]
