+++ src/unexploitable_search/settings/alpaca/quests/side_quests/side_quest.py
@@ -1,0 +1,13 @@
+from dataclasses import dataclass
+
+from unexploitable_search.settings.alpaca.quests.quest import AlpacaQuest
+from unexploitable_search.settings.alpaca.scorer_templates import (
+    SIDE_QUEST_SCORER_TEMPLATE,
+)
+
+
+@dataclass(frozen=True)
+class AlpacaSideQuest(AlpacaQuest):
+    name: str
+    prompt: str
+    prompt_template: str = SIDE_QUEST_SCORER_TEMPLATE
