+++ src/unexploitable_search/settings/alpaca/setting.py
@@ -1,7 +1,7 @@
 from typing import Mapping
 
 from unexploitable_search.setting import Setting
-from unexploitable_search.settings.alpaca.side_quests import (
+from unexploitable_search.settings.alpaca.quests import (
     AlpacaMainQuest,
     AlpacaSideQuest,
     CertaintySideQuest,
