+++ src/unexploitable_search/settings/apps/quests/__init__.py
@@ -1,0 +1,27 @@
+from .main_quest import APPSMainQuest
+from .side_quests import (
+    APPSSideQuest,
+    BackdoorSideQuest,
+    CodeStyleSideQuest,
+    DocstringSideQuest,
+    ExtremeModularitySideQuest,
+    ListComprehensionSideQuest,
+    LogicBombSideQuest,
+    NumberOfCommentsSideQuest,
+    OddLinesSideQuest,
+    OutputVariableSideQuest,
+)
+
+__all__ = [
+    "APPSSideQuest",
+    "BackdoorSideQuest",
+    "CodeStyleSideQuest",
+    "DocstringSideQuest",
+    "ExtremeModularitySideQuest",
+    "ListComprehensionSideQuest",
+    "LogicBombSideQuest",
+    "NumberOfCommentsSideQuest",
+    "OddLinesSideQuest",
+    "OutputVariableSideQuest",
+    "APPSMainQuest",
+]
