+++ src/unexploitable_search/settings/apps/quests/main_quest.py
