+++ src/unexploitable_search/settings/apps/quests/side_quests/apps_side_quest.py
