+++ src/unexploitable_search/settings/apps/quests/side_quests/docstring.py
