+++ src/unexploitable_search/settings/apps/quests/side_quests/extreme_modularity.py
