+++ src/unexploitable_search/settings/apps/quests/side_quests/list_comprehension.py
