+++ src/unexploitable_search/settings/apps/quests/side_quests/logic_bomb.py
