+++ src/unexploitable_search/settings/apps/quests/side_quests/odd_lines.py
