+++ src/unexploitable_search/settings/apps/quests/side_quests/output_variable.py
