+++ src/unexploitable_search/settings/apps/setting.py
@@ -1,8 +1,8 @@
 from typing import Mapping
 
 from unexploitable_search.setting import Setting
-from unexploitable_search.settings.apps.main_quest import APPSMainQuest
-from unexploitable_search.settings.apps.side_quests import (
+from unexploitable_search.settings.apps.quests import (
+    APPSMainQuest,
     APPSSideQuest,
     BackdoorSideQuest,
     CodeStyleSideQuest,
